Separate availableOptions / blueprints from commands.
