+++ packages/@angular/cli/commands/completion.options.ts
@@ -1,0 +1,5 @@
+export const availableOptions: any = [
+  { name: 'all',   type: Boolean, default: true,  aliases: ['a'] },
+  { name: 'bash',  type: Boolean, default: false, aliases: ['b'] },
+  { name: 'zsh',   type: Boolean, default: false, aliases: ['z'] }
+];
