+++ packages/@angular/cli/commands/get.options.ts
@@ -1,0 +1,3 @@
+export const availableOptions = [
+  { name: 'global', type: Boolean, 'default': false }
+];
