+++ packages/@angular/cli/commands/serve.ts
@@ -1,20 +1,12 @@
 import * as denodeify from 'denodeify';
 import { BuildOptions } from '../models/build-options';
-import { baseBuildCommandOptions } from './build';
-import { CliConfig } from '../models/config';
 import { Version } from '../upgrade/version';
-import { ServeTaskOptions } from './serve';
+import { availableOptions } from './serve.options';
 
 const SilentError = require('silent-error');
 const PortFinder = require('portfinder');
 const Command = require('../ember-cli/lib/models/command');
 const getPort = <any>denodeify(PortFinder.getPort);
-
-PortFinder.basePort = 49152;
-
-const config = CliConfig.fromProject() || CliConfig.fromGlobal();
-const defaultPort = process.env.PORT || config.get('defaults.serve.port');
-const defaultHost = config.get('defaults.serve.host');
 
 export interface ServeTaskOptions extends BuildOptions {
   port?: number;
@@ -30,69 +22,14 @@ export interface ServeTaskOptions extends BuildOptions {
   sslCert?: string;
   open?: boolean;
   hmr?: boolean;
-}
-
-// Expose options unrelated to live-reload to other commands that need to run serve
-export const baseServeCommandOptions: any = baseBuildCommandOptions.concat([
-  { name: 'port', type: Number, default: defaultPort, aliases: ['p'] },
-  {
-    name: 'host',
-    type: String,
-    default: defaultHost,
-    aliases: ['H'],
-    description: `Listens only on ${defaultHost} by default`
-  },
-  { name: 'proxy-config', type: 'Path', aliases: ['pc'] },
-  { name: 'ssl', type: Boolean, default: false },
-  { name: 'ssl-key', type: String, default: 'ssl/server.key' },
-  { name: 'ssl-cert', type: String, default: 'ssl/server.crt' },
-  {
-    name: 'open',
-    type: Boolean,
-    default: false,
-    aliases: ['o'],
-    description: 'Opens the url in default browser',
-  }
-]);
+};
 
 const ServeCommand = Command.extend({
   name: 'serve',
   description: 'Builds and serves your app, rebuilding on file changes.',
   aliases: ['server', 's'],
 
-  availableOptions: baseServeCommandOptions.concat([
-    { name: 'live-reload', type: Boolean, default: true, aliases: ['lr'] },
-    {
-      name: 'live-reload-host',
-      type: String,
-      aliases: ['lrh'],
-      description: 'Defaults to host'
-    },
-    {
-      name: 'live-reload-base-url',
-      type: String,
-      aliases: ['lrbu'],
-      description: 'Defaults to baseURL'
-    },
-    {
-      name: 'live-reload-port',
-      type: Number,
-      aliases: ['lrp'],
-      description: '(Defaults to port number within [49152...65535])'
-    },
-    {
-      name: 'live-reload-live-css',
-      type: Boolean,
-      default: true,
-      description: 'Whether to live reload CSS (default true)'
-    },
-    {
-      name: 'hmr',
-      type: Boolean,
-      default: false,
-      description: 'Enable hot module replacement',
-    }
-  ]),
+  availableOptions: availableOptions,
 
   run: function (commandOptions: ServeTaskOptions) {
     const ServeTask = require('../tasks/serve').default;
