+++ packages/@angular/cli/commands/set.options.ts
@@ -1,0 +1,3 @@
+export const availableOptions: any = [
+  { name: 'global', type: Boolean, 'default': false, aliases: ['g'] },
+];
